requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==2.2.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-request==1.5.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
Markdown==3.2.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==1.5
pytz==2020.1
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
